// reference/gateway/src/envelopes.ts
/*
 * Copyright 2025 dgp-standard
 *
 * Licensed under the Apache License, Version 2.0 (the \"License\");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an \"AS IS\" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { GovernanceEnvelope } from "./types";
import { v4 as uuidv4 } from "uuid";

export function getEnvelopeForRequest(system_id: string, action_type: string): GovernanceEnvelope {
  // v0.2: multi-rule envelope with position limits
  return {
    envelope_id: uuidv4(),
    name: "TradingSafetyEnvelope",
    version: "0.2.0",
    scope: {
      system_id,
      action_type
    },
    rules: [
      {
        id: "RULE-MAX-NOTIONAL-001",
        description: "DENY orders with notional > 10_000",
        max_notional: 10000
      },
      {
        id: "RULE-MAX-OPEN-POSITIONS-001",
        description: "DENY new positions when open position count >= 3",
        max_open_positions: 3
      }
    ]
  };
}
